fix: onEnable Webhooks rerender with magic link
