+++ packages/client/src/pages/Webhook.tsx
@@ -5,7 +5,6 @@ import { LoadingButton } from '@mui/lab';
 import { useSvixAccount } from '../context/SvixAccountProvider';
 import { TailSpin } from 'react-loader-spinner';
 import { useApi } from '../data/hooks';
-import { useEnvironment } from '../context/EnvironmentProvider';
 import { useAccount } from '../context/AccountProvider';
 import { SVIX_CONSUMER_APP_PORTAL_URI } from '../constants';
 
@@ -13,12 +12,10 @@ function getSvixConsumerPortalUrl(key) {
     return `${SVIX_CONSUMER_APP_PORTAL_URI}${key}`;
 }
 function Webhook() {
-    const { loading, isExist, setSvixAccount } = useSvixAccount();
-    const { environment } = useEnvironment();
+    const { loading, isExist, handleCreation } = useSvixAccount();
     const { account } = useAccount();
     const { fetch, data } = useApi();
     const [key, setKey] = useState();
-
     useEffect(
         function () {
             async function getMagicLink() {
@@ -46,70 +43,60 @@ function Webhook() {
         [account?.id, fetch, data, account, key, isExist]
     );
 
-    async function handleCreation() {
-        await fetch({
-            method: 'POST',
-            url: `/internal/account/svix`,
-            payload: {
-                environment,
-                accountId: account?.id,
-            },
-        });
-
-        setSvixAccount(data);
-    }
-
     return (
         <div className="w-[80vw]">
+            <MainHeader>
+                <Box
+                    component="div"
+                    sx={{
+                        display: 'flex',
+                        flexDirection: 'column',
+                    }}
+                    className="text-lg mb-4"
+                >
+                    <h1 className="text-3xl font-bold mb-3">Webhooks</h1>
+                    <span className="text-[#b1b8ba]">
+                        {!isExist && !loading
+                            ? 'Enable webhooks to start building and managing'
+                            : 'Build and Manage Webhooks'}
+                    </span>
+                </Box>
+                {!isExist && !loading && (
+                    <Box>
+                        <LoadingButton
+                            variant="text"
+                            style={{
+                                background: '#293347',
+                                color: '#fff',
+                                padding: '0.6rem 1.4rem',
+                                textTransform: 'capitalize',
+                            }}
+                            onClick={() => handleCreation()}
+                        >
+                            Webhook
+                        </LoadingButton>
+                    </Box>
+                )}
+            </MainHeader>
+            {!isExist && !loading && (
+                <div className="flex flex-col justify-center items-center h-[70vh] w-[80vw]">
+                    <p>Enable webhooks to start configure </p>
+                </div>
+            )}
             {loading && (
                 <TailSpin
-                    wrapperStyle={{ justifyContent: 'center', marginTop: '100px' }}
+                    wrapperStyle={{ justifyContent: 'center', marginTop: '28vh' }}
                     color="#1C1C1C"
                     height={80}
                     width={80}
                 />
-            )}
-            {!isExist && !loading && (
-                <>
-                    <MainHeader>
-                        <Box
-                            component="div"
-                            sx={{
-                                display: 'flex',
-                                flexDirection: 'column',
-                            }}
-                            className="text-lg mb-8"
-                        >
-                            <h1 className="text-3xl font-bold mb-3">Webhooks</h1>
-                            <span className="text-[#b1b8ba]">Enable Webhooks</span>
-                        </Box>
-
-                        <Box>
-                            <LoadingButton
-                                variant="text"
-                                style={{
-                                    background: '#293347',
-                                    color: '#fff',
-                                    padding: '0.6rem 1.4rem',
-                                    textTransform: 'capitalize',
-                                }}
-                                onClick={() => handleCreation()}
-                            >
-                                Webhook
-                            </LoadingButton>
-                        </Box>
-                    </MainHeader>
-                    <div className="flex flex-col justify-center items-center h-[70vh] w-[80vw]">
-                        <p>Enable webhooks to start configure </p>
-                    </div>
-                </>
             )}
             {isExist && !loading && key && (
                 <iframe
                     title="svix"
                     id="iframe-svix"
                     src={getSvixConsumerPortalUrl(key)}
-                    style={{ width: '80vw', height: '100%', border: 'none', color: '#fff' }}
+                    style={{ width: '77vw', height: '80vh', border: 'none', color: '#fff', marginLeft: '3vw' }}
                     allow="clipboard-write"
                     loading="lazy"
                 />
